# kafka-admin/.openapi/openapi.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Kafka Admin REST API",
    "description" : "An API to provide REST endpoints for query Kafka for admin operations",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version" : "0.7.1-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Kafka Admin REST API"
  } ],
  "security" : [ {
    "Bearer" : [ ]
  } ],
...
  "paths" : {
    "/rest/topics" : {
      "get" : {
        "tags" : [ "topics" ],
        "summary" : "Retrieves a list of topics",
        "description" : "Returns a list of all of the available topics, or the list of topics that meet the request query parameters. The topics returned are limited to those records the requestor is authorized to view.",
        "operationId" : "getTopics",
        "parameters" : [ {
          "name" : "offset",
          "in" : "query",
          "description" : "Offset of the first record to return, zero-based",
          "schema" : {
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          },
          "deprecated" : true
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "Maximum number of records to return",
          "schema" : {
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "deprecated" : true
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "Number of records per page",
          "schema" : {
            "format" : "int32",
            "default" : "10",
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "name" : "filter",
          "in" : "query",
          "description" : "Filter to apply when returning the list of topics",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Page number",
          "schema" : {
            "format" : "int32",
            "default" : "1",
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "name" : "order",
          "in" : "query",
          "description" : "Order items are sorted",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/SortDirection"
          }
        }, {
          "name" : "orderKey",
          "in" : "query",
          "description" : "Order key to sort the topics by.",
          "required" : false,
          "schema" : {
            "allOf" : [ {
              "$ref" : "#/components/schemas/TopicOrderKey"
            }, {
              "default" : "name"
            } ]
          }
        } ],
        "responses" : {
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/NotAuthorized"
          },
          "500" : {
            "$ref" : "#/components/responses/ServerError"
          },
          "503" : {
            "$ref" : "#/components/responses/ServiceUnavailable"
          },
          "200" : {
            "description" : "List of topics matching the request query parameters. The topics returned are limited to those records the requestor is authorized to view.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TopicsList"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "topics" ],
        "summary" : "Creates a new topic",
        "description" : "Creates a new topic for Kafka.",
        "operationId" : "createTopic",
        "requestBody" : {
          "description" : "Topic to create.",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/NewTopicInput"
              },
              "examples" : {
                "NewTopicExample" : {
                  "$ref" : "#/components/examples/NewTopicExample"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/NotAuthorized"
          },
          "409" : {
            "$ref" : "#/components/responses/Conflict"
          },
          "500" : {
            "$ref" : "#/components/responses/ServerError"
          },
          "503" : {
            "$ref" : "#/components/responses/ServiceUnavailable"
          },
          "201" : {
            "description" : "Topic created successfully.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Topic"
                }
              }
            }
          }
        }
      }
    },
    "/rest/topics/{topicName}" : {
      "get" : {
        "tags" : [ "topics" ],
        "summary" : "Retrieves a single topic",
        "description" : "Topic",
        "operationId" : "getTopic",
        "parameters" : [ {
          "name" : "topicName",
          "in" : "path",
          "description" : "Name of the topic to describe",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "401" : {
            "$ref" : "#/components/responses/NotAuthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "500" : {
            "$ref" : "#/components/responses/ServerError"
          },
          "503" : {
            "$ref" : "#/components/responses/ServiceUnavailable"
          },
          "200" : {
            "description" : "Kafka topic details",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Topic"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "topics" ],
        "summary" : "Deletes a topic",
        "description" : "Deletes the topic with the specified name.",
        "operationId" : "deleteTopic",
        "parameters" : [ {
          "name" : "topicName",
          "in" : "path",
          "description" : "Name of the topic to delete",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Topic deleted successfully."
          },
          "401" : {
            "$ref" : "#/components/responses/NotAuthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "500" : {
            "$ref" : "#/components/responses/ServerError"
          },
          "503" : {
            "$ref" : "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "patch" : {
        "tags" : [ "topics" ],
        "summary" : "Updates a single topic",
        "description" : "Update a the configuration settings for a topic.",
        "operationId" : "updateTopic",
        "parameters" : [ {
          "name" : "topicName",
          "in" : "path",
          "description" : "Name of the topic to update",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TopicSettings"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/NotAuthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "500" : {
            "$ref" : "#/components/responses/ServerError"
          },
          "503" : {
            "$ref" : "#/components/responses/ServiceUnavailable"
          },
          "200" : {
            "description" : "Topic updated successfully.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Topic"
                }
              }
            }
          }
        }
      }
    },
    "/rest/consumer-groups" : {
      "get" : {
        "tags" : [ "groups" ],
        "summary" : "List of consumer groups in the Kafka instance.",
        "description" : "Returns a list of all consumer groups for a particular Kafka instance. The consumer groups returned are limited to those records the requestor is authorized to view.",
        "operationId" : "getConsumerGroups",
        "parameters" : [ {
          "name" : "offset",
          "in" : "query",
          "description" : "Offset of the first record to return, zero-based",
          "schema" : {
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          },
          "deprecated" : true
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "Maximum number of records to return",
          "schema" : {
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "deprecated" : true
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "Number of records per page",
          "schema" : {
            "format" : "int32",
            "default" : "10",
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Page number",
          "schema" : {
            "format" : "int32",
            "default" : "1",
            "minimum" : 1,
            "type" : "integer"
          }
        }, {
          "name" : "topic",
          "in" : "query",
          "description" : "Return consumer groups where the topic name contains with this value",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "group-id-filter",
          "in" : "query",
          "description" : "Return the consumer groups where the ID contains with this value",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "order",
          "in" : "query",
          "description" : "Order items are sorted",
          "required" : false,
          "schema" : {
            "$ref" : "#/components/schemas/SortDirection"
          }
        }, {
          "name" : "orderKey",
          "in" : "query",
          "schema" : {
            "allOf" : [ {
              "$ref" : "#/components/schemas/ConsumerGroupOrderKey"
            }, {
              "default" : "name"
            } ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "List of consumer groups matching the request parameters. The consumer groups returned are limited to those records the requestor is authorized to view.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsumerGroupList"
                },
                "examples" : {
                  "List of consumer groups" : {
                    "value" : {
                      "count" : 1,
                      "limit" : 10,
                      "offset" : 0,
                      "items" : [ {
                        "groupId" : "consumer_group_1",
                        "consumers" : [ {
                          "groupId" : "consumer_group_1",
                          "topic" : "topic-1",
                          "partition" : 0,
                          "memberId" : "consumer_group_member1",
                          "offset" : 5,
                          "lag" : 0,
                          "logEndOffset" : 5
                        }, {
                          "groupId" : "consumer_group_1",
                          "topic" : "topic-1",
                          "partition" : 1,
                          "memberId" : "consumer_group_member2",
                          "offset" : 3,
                          "lag" : 0,
                          "logEndOffset" : 3
                        }, {
                          "groupId" : "consumer_group_1",
                          "topic" : "topic-1",
                          "partition" : 2,
                          "memberId" : "consumer_group_member3",
                          "offset" : 6,
                          "lag" : 1,
                          "logEndOffset" : 5
                        } ]
                      } ]
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/NotAuthorized"
          },
          "500" : {
            "$ref" : "#/components/responses/ServerError"
          },
          "503" : {
            "$ref" : "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
    "/rest/consumer-groups/{consumerGroupId}/reset-offset" : {
      "post" : {
        "tags" : [ "groups" ],
        "summary" : "Reset the offset for a consumer group.",
        "description" : "Reset the offset for a particular consumer group.",
        "operationId" : "resetConsumerGroupOffset",
        "parameters" : [ {
          "name" : "consumerGroupId",
          "in" : "path",
          "description" : "Consumer group identifier",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ConsumerGroupResetOffsetParameters"
              },
              "examples" : {
                "ConsumerGroupOffsetResetExample" : {
                  "$ref" : "#/components/examples/ConsumerGroupOffsetResetExample"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/NotAuthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "$ref" : "#/components/responses/NotFound"
          },
          "500" : {
            "$ref" : "#/components/responses/ServerError"
          },
          "503" : {
            "$ref" : "#/components/responses/ServiceUnavailable"
          },
          "200" : {
            "description" : "The consumer group offsets have been reset.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ConsumerGroupResetOffsetResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/rest/acls/resource-operations" : {
      "get" : {
        "tags" : [ "acls" ],
        "summary" : "Retrieve allowed ACL resources and operations",
        "description" : "Retrieve the resources and associated operations that may have ACLs configured.",
        "operationId" : "getAclResourceOperations",
        "responses" : {
          "200" : {
            "description" : "Map of allowed resources and operations for ACL creation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "array",
                    "items" : {
                      "type" : "string"
                    }
                  }
                },
                "examples" : {
                  "success" : {
                    "summary" : "Map of resources to allowed ACL operations",
                    "value" : "{\n    \"cluster\": [ \"describe\", \"alter\" ],\n    \"group\": [ \"all\", \"delete\", \"describe\", \"read\" ],\n    \"topic\": [ \"all\", \"alter\", \"alter_configs\", \"create\", \"delete\", \"describe\", \"describe_configs\", \"read\", \"write\" ],\n    \"transactional_id\": [ \"all\", \"describe\", \"write\" ]\n}"
                  },
                  "disabled" : {
                    "summary" : "No configuration of ACL bindings is allowed",
                    "value" : "{}"
                  }
                }
              }
            }
          },
          "401" : {
            "$ref" : "#/components/responses/NotAuthorized"
          },
          "500" : {
            "$ref" : "#/components/responses/ServerError"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "SortDirection" : {
        "enum" : [ "asc", "desc" ],
        "type" : "string"
      },
...
      "TopicOrderKey" : {
        "enum" : [ "name", "partitions", "retention.ms", "retention.bytes" ],
        "type" : "string"
      },
...
    },
    "responses" : {
      "BadRequest" : {
        "description" : "The client request was invalid. One or more request parameters or the request\nbody was rejected. Additional information may be found in the response.",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Error"
            },
            "examples" : {
              "badrequest" : {
                "summary" : "Invalid query parameter",
                "value" : {
                  "code" : 400,
                  "error_message" : "Invalid parameter value"
                }
              }
            }
          }
        }
      },
      "NotAuthorized" : {
        "description" : "Request authentication missing or invalid",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ServerError" : {
        "description" : "Internal server error",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Error"
            }
          }
        }
      },
      "ServiceUnavailable" : {
        "description" : "Kafka service unavailable",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "examples" : {
      "NewTopicExample" : {
        "description" : "Sample new topic with 3 partitions",
        "value" : {
          "name" : "my-topic",
          "settings" : {
            "numPartitions" : 3,
            "config" : [ {
              "key" : "min.insync.replicas",
              "value" : "1"
            }, {
              "key" : "max.message.bytes",
              "value" : "1050000"
            } ]
          }
        }
      },
...
    },
    "securitySchemes" : {
      "Bearer" : {
        "type" : "oauth2",
        "description" : "Access token used with OAuth 2.0",
        "flows" : {
          "clientCredentials" : {
            "tokenUrl" : "/token",
            "scopes" : { }
          }
        }
      }
    }
  }
}